### 2018-09-07 #933 @whedon
[PRE REVIEW]: ipc: An R Package for Inter-process Communication
**Submitting author:** @ifellows (<a href="http://orcid.org/0000-0003-1261-0925">Ian Fellows</a>)
**Repository:** <a href="https://github.com/fellstat/ipc" target ="_blank">https://github.com/fellstat/ipc</a>
**Version:** 0.1.0
