### 2018-09-07 #933 @ifellows
note: license specified in package as per https://cran.r-project.org/doc/manuals/r-release/R-exts.html#Licensing
