### 2018-09-07 #933 @yochannah
@ifellows thanks for your suggestions for potential reviewers, that's really handy!

One of the requirements for JOSS papers is that they should have an obvious research application. Could you elaborate a little bit how this package is used for research?

Re: licencing - it would probably be best if you could include a LICENSE file in your repo as well, and add it to your `.Rbuildignore` if you wish.

Thanks!
