### 2018-09-07 #933 @ifellows
Thanks.

Sorry if it wasn't immediately apparent from the paper. Like the recent `datastructures` paper, this package is supporting infrastructure for researchers. It comes out of a CDC initiative to build UIs for complex statistical methods for use by researchers in the field.

Often research requires computationally intensive activity. This package supports HPC by making it easy for the researcher to monitor their distributed processing by having child threads report back partial results and progress. Additionally, shiny is very popular among researchers for building user interfaces to their methods. Currently shiny becomes unresponsive with computationally intensive tasks (see for example http://blog.fellstat.com/?p=407 ). This package fixes this allowing for responsive, scalable applications. Three such applications are included in the package as examples.
