### 2018-09-08 #933 @yochannah
@ifellows Thanks, that does make sense - HPC is definitely an important part of research, and as you say, Shiny is well-loved amongst researchers too - I can see why one might want to make it more able to handle computationally intensive tasks.

Please go ahead and add a LICENSE file to your repo as well. 👍
