### 2018-09-09 #933 @ifellows
@yochannah : License file added. Actually made it MIT while I was at it.
