### 2018-09-10 #933 @JohnCoene
@ifellows @yochannah I had a brief look at the paper, I'm happy to review it but I'm away for the next 2 weeks and can't do anything before 1st October. Feel free to assign someone else if that is too long.
