### 2018-09-10 760bcec
folded vignette into paper
