+++ paper.md
@@ -16,7 +16,8 @@ affiliations:
   name: Fellows Statistics
 ---
 
-# Summary
+Introduction
+-----------
 
 In computer science, asynchronous processing is critical for performing a wide array of tasks, from high performance computing to web services. Communication between these disparate asynchronous processes is often required. Currently the statistical computing language R provides no built in features to handle interprocess communication. Several packages have been written to handle the passing of text or binary data between processes (e.g. [@txtq], [@liteq], and [@rzmq]).
 
@@ -36,6 +37,405 @@ q$consumer$consume()
 ```
 
 This package can be applied to high performance computing environments, easily allowing parallel worker processes to communicate partial results or progress to the main thread. Another major use case is to support asynchronous web based user interfaces ([@shiny]) to long running statistical algorithms.
+
+
+Quick Start
+-----------
+
+Perhaps the most important use of this package is to communicate with a
+parent thread from a child thread executed using the `future` package.
+This is very easy to do. Simply create and start a `Queue` in the parent
+thread. The child thread can then send messages and evaluate R code on
+the main thread.
+
+    library(ipc)
+    library(future)
+    plan(multiprocess)
+    q <- queue()
+
+    value <- ""
+
+    f <- future({
+      # set value in the main thread
+      q$producer$fireEval({
+        value <- "Hello world"
+      })
+    })
+
+    Sys.sleep(.5)
+    # Evaluate signals
+    cons <- q$consumer$consume()
+    print(value)
+
+    ## [1] "Hello world"
+
+    # Remove temporary files
+    q$destroy()
+
+A `Queue` object has a `producer` field that is used to send signals
+onto the `Queue` and a `consumer` field that is used to read from the
+queue, and process any signals written to the `Queue`.
+
+Messages can be consumed by calling the consumer's `consume` method and
+will be handled by default using the environment that consume is called
+in. The consumer's `start` method will execute consume at regular
+intervals provided the R process is idle.
+
+By default, a `Consumer` object knows how to handle two signals. The
+`"eval"` handler will execute the signal's data as an expression in the
+evaluation environment (i.e. where `consume` is called. The `"doCall"`
+handler will call the function defined by the first element of the data
+with parameters equal to the second element. Functions to handle various
+signals can be added to the consumer using the `addHandler` method.
+
+`Producer` objects has a built in method to signal for evaluation.
+`fireEval(expr, env)` will signal for `expr` to be evaluated,
+substituting in any value in `env`. For example, the following code will
+set the variable val to 2 in the environment in which it is consumed.
+
+    variable <- 2
+    q$producer$fireEval(val <- j, env=list(j=variable))
+
+### Signaling a child process
+
+Signals can also be sent from the main thread to a child. Here we will
+cause the child process to throw an error.
+
+    library(future)
+    library(promises)
+    plan(multiprocess)
+
+    q <- queue()
+
+    fut <- future({
+      for(i in 1:1000){
+        Sys.sleep(.1)
+        q$consumer$consume()
+     }
+    })
+
+    q$producer$fireEval(stop("Stop that child"))
+    cat(try(value(fut)))
+
+    ## Error in eval(obj, envir = env) : Stop that child
+
+    q$destroy()
+
+If errors occur during the consumption, all messages are processed, and
+then the first error encountered is then thrown. Alternatively, errors
+can be switched to warnings using
+`q$consumer$consume(throwErrors=FALSE)`.
+
+### Continuous consumption
+
+A consumer's `start` method can be used to execute `comsume` at regular
+intervals (provided the R process is idle).
+
+    library(future)
+    library(promises)
+    plan(multiprocess)
+
+    q <- queue()
+
+    fut <- future({
[…]
