### 2018-09-10 f88a9aa
Cran submission
