### 2018-09-10 72f897f
I guess R CMD check needs the implicit future cluster to be cleaned up explicitly in examples (but not tests).
