### 2018-09-10 1aa7958
edits for cran
