### 2018-09-13 #933 @mschubert
As the author of the `txtq` package (which this package builds on), @wlandau might have some good input?
