### 2018-09-13 #933 @wlandau
Thanks for the endorsement. However, since [`ipc`](https://github.com/fellstat/ipc) builds on [`txtq`](https://github.com/wlandau/txtq), I am not sure a review from me would be sufficiently free of bias. (And like @JohnCoene, I will be away from Sept 15 through Oct 1). I think @clarkfitzg would be an excellent second reviewer, should he accept.
