### 2018-09-30 #933 @JohnCoene
@yochannah Just got back; yes, I can help review this paper.
