### 2018-10-01 fellstat/ipc#1 @JohnCoene
[JOSS-review] Paper - Documentation
A few things to kicks things off. I think the paper includes too much documentation, as stated in the guidelines:

> the paper should not include software documentation such as API (Application Programming Interface) functionality, as this should be outlined in the software documentation.

The examples in the paper are already included in the [introductory vignette](http://htmlpreview.github.io/?https://github.com/fellstat/ipc/blob/master/inst/doc/shinymp.html).

related to https://github.com/openjournals/joss-reviews/issues/988
