### 2018-10-01 fellstat/ipc#2 @JohnCoene
[JOSS-review] LICENSE - Content
The `LICENSE` file needs to include content.

related to https://github.com/openjournals/joss-reviews/issues/988
