### 2018-10-01 fellstat/ipc#3 @JohnCoene
[JOSS-review] Statement of Need
The Statement of Need is missing from the `README`, you already have a good one in the paper, I assume you could use that.

related to https://github.com/openjournals/joss-reviews/issues/988
