### 2018-10-01 fellstat/ipc#4 @JohnCoene
[JOSS-review] Contribution Guidelines
The `README` is missing Countribution Guidelines:

> There should be clear guidelines for third-parties wishing to:
>
> - Contribute to the software
> - Report issues or problems with the software
> - Seek support

related to https://github.com/openjournals/joss-reviews/issues/988
