### 2018-10-01 fellstat/ipc#5 @JohnCoene
[JOSS-review] Installation
There is external software required for installation, unless I am mistaken. A new user, like myself, blindly running `install.packages` or `remotes::install_github` will run into an error, namely:

```r
Configuration failed because hiredis was not found. Try installing:
 * deb: libhiredis-dev (Debian, Ubuntu, etc)
 * rpm: hiredis-devel (Fedora, EPEL)
 * brew: hiredis (OSX)
```

This should probably be mentioned in the installation instruction.

EDIT: related to [JOSS review](https://github.com/openjournals/joss-reviews/issues/988)
