### 2018-10-01 fellstat/ipc#5 @ifellows
Can you please include the full output of the installation attempt, along with `sessionInfo()`.  If redis is causing any installation issues I'll move it to a Suggest, as it is not the default storage backend.
