### 2018-10-01 fellstat/ipc#5 @JohnCoene
Install

```r
devtools::install_github("fellstat/ipc")
Downloading GitHub repo fellstat/ipc@master
from URL https://api.github.com/repos/fellstat/ipc/zipball/master
Installing ipc
trying URL 'https://cloud.r-project.org/src/contrib/redux_1.1.0.tar.gz'
Content type 'application/x-gzip' length 76920 bytes (75 KB)
==================================================
downloaded 75 KB

Installing redux
'/usr/lib/R/bin/R' --no-site-file --no-environ --no-save --no-restore --quiet CMD INSTALL  \
  '/tmp/RtmpgHJAYC/devtoolsce7771652bf/redux'  \
  --library='/home/jp/R/x86_64-pc-linux-gnu-library/3.4' --install-tests

* installing *source* package ‘redux’ ...
** package ‘redux’ successfully unpacked and MD5 sums checked
Package hiredis was not found in the pkg-config search path.
Perhaps you should add the directory containing `hiredis.pc'
to the PKG_CONFIG_PATH environment variable
No package 'hiredis' found
Using PKG_CFLAGS=-I/usr/include/hiredis
Using PKG_LIBS=-lhiredis
------------------------- ANTICONF ERROR ---------------------------
Configuration failed because hiredis was not found. Try installing:
 * deb: libhiredis-dev (Debian, Ubuntu, etc)
 * rpm: hiredis-devel (Fedora, EPEL)
 * brew: hiredis (OSX)
If hiredis is already installed, check that 'pkg-config' is in your
PATH and PKG_CONFIG_PATH contains a hiredis.pc file. If pkg-config
is unavailable you can set INCLUDE_DIR and LIB_DIR manually via:
R CMD INSTALL --configure-vars='INCLUDE_DIR=... LIB_DIR=...'
--------------------------------------------------------------------
ERROR: configuration failed for package ‘redux’
* removing ‘/home/jp/R/x86_64-pc-linux-gnu-library/3.4/redux’
Installation failed: Command failed (1)
'/usr/lib/R/bin/R' --no-site-file --no-environ --no-save --no-restore --quiet CMD INSTALL  \
  '/tmp/RtmpgHJAYC/devtoolsce7ffdfbe8/fellstat-ipc-1aa7958'  \
  --library='/home/jp/R/x86_64-pc-linux-gnu-library/3.4' --install-tests

ERROR: dependency ‘redux’ is not available for package ‘ipc’
* removing ‘/home/jp/R/x86_64-pc-linux-gnu-library/3.4/ipc’
Installation failed: Command failed (1)
```

Session

```r
sessionInfo()
R version 3.4.4 (2018-03-15)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Ubuntu 18.04.1 LTS

Matrix products: default
BLAS: /usr/lib/x86_64-linux-gnu/blas/libblas.so.3.7.1
LAPACK: /usr/lib/x86_64-linux-gnu/lapack/liblapack.so.3.7.1

locale:
 [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C               LC_TIME=en_GB.UTF-8
 [4] LC_COLLATE=en_US.UTF-8     LC_MONETARY=en_GB.UTF-8    LC_MESSAGES=en_US.UTF-8
 [7] LC_PAPER=en_GB.UTF-8       LC_NAME=C                  LC_ADDRESS=C
[10] LC_TELEPHONE=C             LC_MEASUREMENT=en_GB.UTF-8 LC_IDENTIFICATION=C

attached base packages:

other attached packages:
[1] future_1.9.0   ipc_0.1.0      testthat_2.0.0

 [1] Rcpp_0.12.18     compiler_3.4.4   later_0.7.4      git2r_0.23.0     tools_3.4.4
 [6] digest_0.6.16    memoise_1.1.0    rlang_0.2.2      cli_1.0.0        shiny_1.1.0
[11] rstudioapi_0.7   filelock_1.0.1   commonmark_1.5   parallel_3.4.4   curl_3.2
[16] yaml_2.2.0       withr_2.1.2      stringr_1.3.1    httr_1.3.1       roxygen2_6.1.0
[21] xml2_1.2.0       knitr_1.20       globals_0.12.1   desc_1.2.0       fs_1.2.6
[26] devtools_1.13.6  rprojroot_1.3-2  listenv_0.7.0    R6_2.2.2         redux_1.1.0
[31] base64url_1.4    txtq_0.1.0       purrr_0.2.5      magrittr_1.5     codetools_0.2-15
[36] backports_1.1.2  promises_1.0.1   htmltools_0.3.6  assertthat_0.2.0 mime_0.5
[41] xtable_1.8-2     httpuv_1.4.5     stringi_1.2.4    crayon_1.3.4
```
