### 2018-10-01 fellstat/ipc#5 @ifellows
Thank you, it appears that redux will not install from source without a working redis installation. On Mac and windows, `ipc` installs fine because of the binary distributions of `redux` present on CRAN. This is enough of an issue to warrant sidelining the redis backend to a Suggest.
