### 2018-10-01 fellstat/ipc#6 @JohnCoene
[JOSS-review] Tests - coverage
I will check whether this is necessary but though the package includes unit tests which all pass when cloned and tested, because there is no continuous integration (i.e. [codecov](https://codecov.io/) or [coveralls](https://coveralls.io/)) it is not possible for me to know how much of the code is tested.

related to https://github.com/openjournals/joss-reviews/issues/988
