### 2018-10-01 #988 @JohnCoene
@yochannah Unit tests are present in the package's source code but not coverage is reported due to lack of continuous integration (codecov/coveralls), is it necessary to require adding that?
