### 2018-10-01 fellstat/ipc#6 @ifellows
Code coverage is not currently computed during CI. Continuous integration is done via Travis

https://travis-ci.org/fellstat/ipc

FYI, here is current test coverage:

```
> covr::package_coverage(type = "tests")
ipc Coverage: 39.66%
R/async-interrupt.R: 0.00%
R/async-progress.R: 0.00%
R/example.R: 0.00%
R/queue.R: 0.00%
R/redis-source.R: 92.59%
R/globals.R: 16.67%
R/consumer.R: 61.29%
R/producer.R: 61.54%
R/text-file-source.R: 92.59%
```
