### 2018-10-01 #988 @yochannah
@JohnCoene I think CI goes under the "optional but strongly recommended" category - https://joss.readthedocs.io/en/latest/review_criteria.html#tests - adding CI would be awesome but given we have automated tests that is probably sufficient.
