### 2018-10-01 2f098cd
Added statement of need to README. issue #3
