### 2018-10-01 325a128
Added full MIT license to LICENCE file. issue #2
