### 2018-10-01 410add1
Added link to contribution guidelines in readme. issue #4
