### 2018-10-01 f86da8b
pared down paper length. issue #1
