### 2018-10-01 fellstat/ipc#2 @ifellows
CRAN is _very particular_ with regard to LICENSE file formats. I've added a LICENCE file with the full MIT license.
