### 2018-10-01 fellstat/ipc#5 @ifellows
redux is now a suggest. Thanks for this. This wold have likely been a stumbling point for people.
