### 2018-10-02 fellstat/ipc#2 @mschubert
What's the issue with having one LICENSE file recognized by CRAN *and* Github?
