### 2018-10-02 fellstat/ipc#6 @mschubert
FWIW, the way I interpret the JOSS guidelines is that automated test coverage calculation is not required. In this case, it would fall to the reviewers (*us*) to see if the tests present are sufficient.
