### 2018-10-02 fellstat/ipc#2 @ifellows
For CRAN: The MIT license is a template, with the date and author needing to be specified. How they want this done is with "MIT + file LICENSE" in the description and the template parameters (and only the template parameters) specified in the LICENSE file.

Sometimes you just have to jump through the hoops they put in front of you.
