### 2018-10-03 fellstat/ipc#2 @JohnCoene
This could be because there are now two `LICENSE` file.
