### 2018-10-06 fellstat/ipc#8 @mschubert
[JOSS-review] Linking of relevant information
related: https://github.com/openjournals/joss-reviews/issues/988

I think the following adjustments would improve accessibility. Most of them are not strict requirements, but if unresolved I would list this for the JOSS editor to comment upon.

* [x] Paper affiliation doesn't mean anything to me. I guess it's a company? If so, use the full name, some location information, and potentially link to the web page
* [x] It would be nice if the README links to CRAN (e.g. using a [CRAN badge](https://www.r-pkg.org/services#badges)), and CRAN links to the Github page (using `URL:` and `BugReports:` fields in `DESCRIPTION`)
* [ ] The `LICENSE` file should be machine readable for both CRAN and Github (related: #2)
* [x] The contributing guidelines can be improved: This is currently a rather long wiki document which describes git/Github basics. This is great for people who have never used it, but for others shorter information<sup>1</sup> could be put the in the `README` directly, with a link to the long document for more details. Possibly (but not required) use Issue/PR templates.
In particular, also:
  * [x] The current contribution guidelines put manual testing workload on contributors that could be solved using CI (e.g., Travis builds and tests pull requests automatically - this should only take an hour to set up)
  * [x] Advice in [How to contribute](https://github.com/fellstat/ipc/wiki/How-to-Contribute:-Git-Practices) is mostly internal guidelines, which bury relevant information for the wider community at the bottom. Doesn't it make more sense to have succinct instructions for external contributors linked here? (and you can always put the internal ones a separate document). JOSS requires *guidelines for third parties* (related: #4).
  * [x] External advice about pulling the latest master is not necessary as long as there are no conflicts (Github PR merge interface supports rebasing). If there are conflicts, the comments do not help resolving them.
  * [x] Support states to contact the author, but not how. Using issues, email, etc.? (Issues would have the advantage to be publicly searchable)

<sup>1</sup> e.g. "Please report issues through the [Github issues page](https://github.com/fellstat/ipc/issues) [...]"
