### 2018-10-06 fellstat/ipc#9 @mschubert
[JOSS-review] Paper and package claims
related: https://github.com/openjournals/joss-reviews/issues/988

I think the use of this package to develop responsive applications is very well supported, and as far as I can tell this was the reason this package was developed in the first place (commit history also lists the previous name as `ShinyAsyncTools`). It is definitely useful as such, and the `shiny` and `future` examples using the progress bar instead of the application becoming unresponsive illustrate this nicely.

However, it looks to me like the paper introduction and statement of need can be improved, to focus more on the packages strengths and not over-claim its novelty. In particular:

* [x] *Currently the statistical computing language R provides no built in features to handle interprocess communication.*
What about the `parallel` or `snow` packages? They do provide communication between processes. `parallel:::sendCall` is also able to send individual instructions to workers (it is not exported, but it exists)
* [x] *What is lacking is a framework to easily pass R objects between processes along with an associated signal, and have handler functions automatically execute them in the receiving process.*
Doesn't [`rmote`](https://github.com/cloudyr/rmote) do this?
* [x] In the [pre-review thread](https://github.com/openjournals/joss-reviews/issues/933#issuecomment-419578414), you stated *Often research requires computationally intensive activity. This package supports HPC by making it easy for the researcher to monitor their distributed processing by having child threads report back partial results and progress.*
What is the advantage over using an actual HPC workflow package here, like `drake`? It does distributed processing, progress report, and that without the user having to write their own signals for it.
If this package is suitable for tracking HPC progress as it claims, I would also like to see a performance evaluation (e.g., how many inputs can it process per time interval?).
