### 2018-10-07 fellstat/ipc#9 @ifellows
Thanks for this. The impetus for this package came directly from shiny, however, what came out was quite a bit more general. The name change came at the suggestion of some folks over at rstudio, who felt the package was being done a disservice by the narrow name.

Clearly a I need to give a bit more thought to how I can communicate the non-shiny use cases. Currently there aren't any examples, so it is sort of left to the imagination. I'll add an example to the vignette, and review the paper to make sure the claims are a bit more narrow.

Fundamentally the difference with some of the things you've mentioned is that this package is focused on communication during computation, and in particular reporting information back to the user in an interactive setting. `parallel:::sendCall` is different because it creates a new computation and sends it to be executed, this is different than communicating with an already running computation. `drake` is a full fledged pipeline, whereas `ipc` operates a bit more like a tool that would be used within a pipeline. `rmote` seems (at first glance at least) to focus on forwarding graphical output from a sever to a client.

One use case where I see this package fitting in with regard to HPC is supporting interactive diagnostics and progress for parallelized functions. For example `ergm`, `rstan` and my own package [lolog](https://github.com/statnet/lolog) all have (optional) parallel operations. It would be great if the running diagnostics functioned identically to how they do when the algorithms are run sequentially. `ergm` and `lolog` drop key diagnostics and progress, while `rstan` can optionally emit to a file that you can tail.

Here is a minimal example of how `ipc` can be used to monitor the progress and convergence of parallel MCMC chains:

```
library(parallel)
library(ipc)

mcmcTask <- function(){
  on.exit(q$destroy())
  q <- queue()
  prog <- 1
  pb <- txtProgressBar(min = 0, max = 40, style=3)
  res <- list()
  chains <- list()
  for(i in 1:4){
    res[[i]] <- mcparallel({
      chain <- c()
      for(j in 1:10){
        chain <- c(chain, rnorm(100))
        q$producer$fireEval(chains[[i]] <- chain, list(i=i,chain=chain))
        q$producer$fireEval(prog <- prog + 1)
        Sys.sleep(runif(1)*5)
      }
      chain
    })
  }
  while(prog < 40){
    q$consumer$consume()
    setTxtProgressBar(pb, prog)
    par(mfrow=c(2,2))
    for(j in seq_along(chains))
      if(length(chains[[j]]) > 1)
        plot(chains[[j]], main=paste("Trace Plot for MCMC chain", j))
    Sys.sleep(1)
  }
  close(pb)
  mccollect(res)
}

chains <- mcmcTask()
```

I'll definitely put some changes into the documentation around this, and I think that it bears some thinking as to whether there are some package changes that would be desirable. More to come....
