### 2018-10-18 e60d683
Added additional example to vignette. switched licence files. edits to paper
