### 2018-10-18 081514a
doc changes
