+++ DESCRIPTION
@@ -8,6 +8,8 @@ Description: Provides tools for passing messages between R processes.
     Shiny Examples are provided showing how to perform useful tasks such as: 
     updating reactive values from within a future, progress bars for long running 
     async tasks, and interrupting async tasks based on user input.
+URL: https://github.com/fellstat/ipc
+BugReports: https://github.com/fellstat/ipc/issues
 Imports: R6, shiny, txtq
 License:  MIT + file LICENCE
 Encoding: UTF-8
