### 2018-10-18 e9e1566
doc changes
