### 2018-10-18 fellstat/ipc#9 @ifellows
So after some thought, what I'd _like_ to do is see if I can wrap parallel clusters with an associated queue and then have messages automatically consumed as you are running any or the various parallel applys. I think this is possible after some looking into the code.

However, it is beyond the current scope. I've worked over the language in the paper, vignette and readme to make it softer with regard to HPC. I've also added an HPC example to the vignette.

Do these changes and the clarifications in my previous comment adequately address your concerns?
