### 2018-10-18 fellstat/ipc#8 @ifellows
> related: [openjournals/joss-reviews#988](https://github.com/openjournals/joss-reviews/issues/988)
>
> I think the following adjustments would improve accessibility. Most of them are not strict requirements, but if unresolved I would list this for the JOSS editor to comment upon.
>
> * [ ]  Paper affiliation doesn't mean anything to me. I guess it's a company? If so, use the full name, some location information, and potentially link to the web page

Added URL.

> * [ ]  It would be nice if the README links to CRAN (e.g. using a [CRAN badge](https://www.r-pkg.org/services#badges)), and CRAN links to the Github page (using `URL:` and `BugReports:` fields in `DESCRIPTION`)

Added link.

> * [ ]  The `LICENSE` file should be machine readable for both CRAN and Github (related: #2)

Ugg. Finally able to resolve this. CRAN looks at LICENCE and github looks at LICENSE.

> * [ ]  The contributing guidelines can be improved: This is currently a rather long wiki document which describes git/Github basics. This is great for people who have never used it, but for others shorter information1 could be put the in the `README` directly, with a link to the long document for more details. Possibly (but not required) use Issue/PR templates.

Given past experience, I've found having all this explicit to be useful for people, especially in the R community where programming is a secondary concern for many people.

>   In particular, also:
>
>   * [ ]  The current contribution guidelines put manual testing workload on contributors that could be solved using CI (e.g., Travis builds and tests pull requests automatically - this should only take an hour to set up)

Travis does automatically test pull requests. However, the contributor should run check prior to filing the pull request. Otherwise they would be firing off a pull request without knowing if the code was in a working state. Of course this would be caught by CI, but it isn't good practice.

>   * [ ]  Advice in [How to contribute](https://github.com/fellstat/ipc/wiki/How-to-Contribute:-Git-Practices) is mostly internal guidelines, which bury relevant information for the wider community at the bottom. Doesn't it make more sense to have succinct instructions for external contributors linked here? (and you can always put the internal ones a separate document). JOSS requires _guidelines for third parties_ (related: #4).

I've moved all the info for external contributors to the top. If you know all about git, no need to read any further.

>   * [ ]  External advice about pulling the latest master is not necessary as long as there are no conflicts (Github PR merge interface supports rebasing). If there are conflicts, the comments do not help resolving them.

I disagree. merge compatibility only assures that the texts are compatible, not the actual code. The guidelines state that you should run check before the pull request. This isn't useful unless the contributor is up-to-date with master.

>   * [ ]  Support states to contact the author, but not how. Using issues, email, etc.? (Issues would have the advantage to be publicly searchable)

added.

>
> 1 e.g. "Please report issues through the [Github issues page](https://github.com/fellstat/ipc/issues) [...]"
