### 2018-10-18 fellstat/ipc#8 @ifellows
please let me know if these changes sufficiently address your comments.
