### 2018-10-21 fellstat/ipc#8 @mschubert
Almost there!

> > * [x] Paper affiliation doesn't mean anything to me. I guess it's a company? If so, use the full name, some location information, and potentially link to the web page
>
> Added URL.

I don't see it in the [compiled paper](https://github.com/openjournals/joss-papers/blob/joss.00988/joss.00988/10.21105.joss.00988.pdf)?

> > * [x]   The `LICENSE` file should be machine readable for both CRAN and Github (related: #2)
>
> Ugg. Finally able to resolve this. CRAN looks at LICENCE and github looks at LICENSE.

Github still doesn't recognize it (but does link to the right file now). Fixing not required for JOSS, it just would be nice if this worked.
