### 2018-10-21 fellstat/ipc#9 @mschubert
I still don't agree with the sentence

> Currently the statistical computing language R provides no built in features to handle interprocess communication.

What `parallel:::sendCall` does is not the same as `ipc` does, but it *is a* form of communication between active R processes as well (the communication being the call sent).

I'd be happy with replacing *interprocess communication* with *reactive values*, or *status values*, or something else you would think fits (*updating status while running computations*?).

<hr/>
Also, it would be great if your statement of need in the paper included your explanations above. This could be as easy as copy + paste the lines

> the difference <with some of the things you've mentioned> is that this package is focused on communication during computation, and in particular reporting information back to the user in an interactive setting. parallel:::sendCall is different because it creates a new computation and sends it to be executed, this is different than communicating with an already running computation. drake is a full fledged pipeline, whereas ipc operates a bit more like a tool that would be used within a pipeline. rmote seems (at first glance at least) to focus on forwarding graphical output from a sever to a client

and

> One use case where I see this package fitting in with regard to HPC is supporting interactive diagnostics and progress for parallelized functions.
