+++ paper.bib
@@ -5,7 +5,7 @@
     note = {R package version 1.1.0},
     url = {https://CRAN.R-project.org/package=shiny},
   }
-  
+
   @Manual{txtq,
     title = {txtq: A Small Message Queue for Parallel Processes},
     author = {William Michael Landau},
@@ -28,4 +28,21 @@
     year = {2017},
     note = {R package version 0.9.3},
     url = {https://CRAN.R-project.org/package=rzmq},
+  }
+
+  @Manual{parallel,
+    title = {R: A Language and Environment for Statistical Computing},
+    author = {{R Core Team}},
+    organization = {R Foundation for Statistical Computing},
+    address = {Vienna, Austria},
+    year = {2018},
+    url = {https://www.R-project.org/},
+  }
+
+  @Manual{drake,
+    title = {drake: A Pipeline Toolkit for Reproducible Computation at Scale},
+    author = {William Michael Landau},
+    year = {2018},
+    note = {R package version 6.0.0},
+    url = {https://CRAN.R-project.org/package=drake},
   }
