+++ paper.md
@@ -19,7 +19,7 @@ affiliations:
 Summary
 -----------
 
-Asynchronous processing is critical for performing a wide array of tasks, from high performance computing to web services. Communication between these disparate asynchronous processes is often required. Currently the statistical computing language R provides no built in features to handle interprocess communication between R processes while they are performing computations. The `parallel` package, which is part of the R computing environment [@parallel] allows the user to send computations tobe executed by idle worker processes, and the `drake` package [@drake] is a general purpose workflow manager, where tasks can be executed in parallel. Several packages have been written to handle the passing of text or binary data between processes (e.g. [@txtq], [@liteq], and [@rzmq]). `ipc` allows you to easily pass R objects between processes along with an associated signal, and have handler functions automatically execute them in the receiving process. Additionally, it is desirable to have a system that can be backed flexibly either through the file system or a database connection. The `ipc` R package aims to fill this need.
+Asynchronous processing is critical for performing a wide array of tasks, from high performance computing to web services. Communication between these disparate asynchronous processes is often required. Currently the statistical computing language R provides no built in features to handle interprocess communication between R processes while they are performing computations. The `parallel` package, which is part of the R computing environment [@parallel] allows the user to send computations to be executed by idle worker processes, and the `drake` package [@drake] is a general purpose workflow manager, where tasks can be executed in parallel. Several packages have been written to handle the passing of text or binary data between processes (e.g. [@txtq], [@liteq], and [@rzmq]). `ipc` allows you to easily pass R objects between processes along with an associated signal, and have handler functions automatically execute them in the receiving process. Additionally, it is desirable to have a system that can be backed flexibly either through the file system or a database connection. The `ipc` R package aims to fill this need.
 
 For example, one might signal for the execution of an expression in one thread to set a variable `a`.
 
@@ -34,7 +34,7 @@ Then in another thread, this signal can be processed, resulting in the value `a`
 q$consumer$consume()
 ```
 
-This package can be applied to high performance computing environments, easily allowing parallel worker processes to communicate partial results or progress to the main thread; however, particular focus is paid to the use case of supporting asynchronous web based user interfaces ([@shiny]).
+This package can be applied to high performance computing environments, easily allowing parallel worker processes to communicate partial results or progress to the main thread. This functionality can be used to supporting interactive diagnostics and progress for parallelized functions. Particular focus is paid to the use case of supporting asynchronous web based user interfaces ([@shiny]).
 
 ### Use in Shiny
 
