### 2018-10-21 fellstat/ipc#9 @ifellows
That is why I had changed the sentence to add "between active R processes," to indicate that the use case is for communication between processes that are in the middle of running R computations. To avoid ambiguity, the sentence is now
> Currently the statistical computing language R provides no built in features to handle interprocess communication between R processes while they are performing computations.

Further, `sendCall` is not an exported function, and therefore could not be said to have been "provided." Not that this matters though, since any parallel functionality *could* be said to be doing interprocess communication. After all, how could you run anything in parallel if you can't communicate what to run?

-----------------------------------

I've added references to `drake` and `parallel` to help contextualize where the package lays within the R ecosystem. I didn't particularly call out `sendCall` as it is not exported and therefore would likely just confuse the reader.

The last sentence before "## Use in Shiny" lays out the HPC use case as requested.
