### 2018-10-22 fellstat/ipc#9 @mschubert
> That is why I had changed the sentence to add "between active R processes,"

Copy/paste error on my part, sorry. My point referred to your updated sentence (as the worker will be started separately, so the worker is running in both cases).

I agree with "provided", but would also argue that if the meaning of your sentence relies on a single word being interpreted correctly, you might want to rephrase in avoid confusion. I would consider myself more knowledgeable than the average user about this topic, and I was confused. How about a user that doesn't know `ipc`, the inner workings of `parallel`, `rzmq`, or `drake`? How should they know if they want to use `ipc` or rather another package? (maybe wait a day, and then re-read your first paper paragraph again - you still may want to clean it up a bit)

> Not that this matters though, since any parallel functionality could be said to be doing interprocess communication. After all, how could you run anything in parallel if you can't communicate what to run?

Yes, that was point of my criticism about positioning `ipc` as *the* way to do inter-process communication.

<hr>

Not mentioning `sendCall` is perfectly fine, as long as the text doesn't pretend the functionality does not exist.

HPC use case looks good.
