### 2018-10-26 65a838f
Added promises to ignette for #10 and some minor paper edits for #9
