+++ paper.md
@@ -19,7 +19,7 @@ affiliations:
 Summary
 -----------
 
-Asynchronous processing is critical for performing a wide array of tasks, from high performance computing to web services. Communication between these disparate asynchronous processes is often required. Currently the statistical computing language R provides no built in features to handle interprocess communication between R processes while they are performing computations. The `parallel` package, which is part of the R computing environment [@parallel] allows the user to send computations to be executed by idle worker processes, and the `drake` package [@drake] is a general purpose workflow manager, where tasks can be executed in parallel. Several packages have been written to handle the passing of text or binary data between processes (e.g. [@txtq], [@liteq], and [@rzmq]). `ipc` allows you to easily pass R objects between processes along with an associated signal, and have handler functions automatically execute them in the receiving process. Additionally, it is desirable to have a system that can be backed flexibly either through the file system or a database connection. The `ipc` R package aims to fill this need.
+Asynchronous processing is critical for performing a wide array of tasks, from high performance computing to web services. Communication between these disparate asynchronous processes is often required. The `parallel` package, which is part of the R computing environment [@parallel] allows the user to send computations to be executed by idle worker processes, and the `drake` package [@drake] is a general purpose workflow manager, where tasks can be executed in parallel. Several packages have been written to handle the passing of text or binary data between processes (e.g. [@txtq], [@liteq], and [@rzmq]).  Currently the statistical computing language R provides no built in features to handle interprocess communication between R processes while they are performing computations. `ipc` allows you to easily pass R objects between processes along with an associated signal, and have handler functions automatically execute them in the receiving process. In `ipc`, communication is backed either through the file system or a database connection.
 
 For example, one might signal for the execution of an expression in one thread to set a variable `a`.
 
