+++ vignettes/shinymp.Rmd
@@ -192,6 +192,7 @@ shinyApp(ui = ui, server = server)
 library(shiny)
 library(ipc)
 library(future)
+library(promises)
 plan(multiprocess)
 
 ui <- fluidPage(
