### 2018-10-26 #988 @ifellows
@mschubert , @JohnCoene: I believe that comments have been addressed. Does everything look good to you both in terms of moving the paper forward?
