### 2018-10-30 #988 @ifellows
@JohnCoene: Does everything look good in terms of moving the paper forward?
