### 2018-11-04 #988 @yochannah
@ifellows - At this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive?
