### 2018-11-05 1d2d5a1
added txtq import to namespace
