+++ NAMESPACE
@@ -17,4 +17,5 @@ export(shinyExample)
 export(shinyQueue)
 export(stopMulticoreFuture)
 export(tempFileGenerator)
+import(txtq)
 importFrom(R6,R6Class)
