### 2018-11-05 #988 @ifellows
10.5281/zenodo.1478139

fyi: revised package on it's way to CRAN
