### 2018-11-05 #988 @ifellows
A hearty thank you to @JohnCoene and @mschubert for reviewing and to @yochannah for shepherding the process. The package has benefited significantly from all of your attention.
