### 2018-11-06 #988 @arfon
@mschubert , @JohnCoene - many thanks for your reviews here and to @yochannah for editing this submission ✨

@ifellows - your paper is now accepted into JOSS :zap: :rocket: :boom:
